Added Ajax delete

+++ app/controllers/ExerciseController.php
@@ -4,7 +4,6 @@ class ExerciseController extends BaseController
 {
 	public function __construct() {
 
-		# Make sure BaseController construct gets called
 		parent::__construct();
 			
 	}
@@ -28,16 +27,12 @@ class ExerciseController extends BaseController
     public function postAdd()
     {
 
-		$rules = array(
-			'title' => 'required'
-		);
-
+		$rules = array('title' => 'required');
 
 		$validator = Validator::make(Input::all(), $rules);
 
-
-		if($validator->fails()) {
-
+		if($validator->fails()) 
+		{
 			return Redirect::to('/exercise/add')
 				->with('flash_message', 'Add failed; please try again.')
 				->withInput()
@@ -48,10 +43,12 @@ class ExerciseController extends BaseController
 		$exercise->desc = filter_var(Input::get('title'), FILTER_SANITIZE_STRING);
 		$exercise->user()->associate(Auth::user()); 
 		
-		try {
+		try 
+		{
 			$exercise->save();
 		}
-		catch (Exception $e) {
+		catch (Exception $e) 
+		{
 			
 			$errorvalue = Helper::getErrorMessage($e);
 
@@ -60,7 +57,6 @@ class ExerciseController extends BaseController
 				->withInput()
 				->withErrors($errorvalue);
 		}
-
 		
 		return Redirect::to('/exercise/index');
 
@@ -68,17 +64,20 @@ class ExerciseController extends BaseController
     
     public function getEdit($id)
     {
-       try{ 
-       $exercise = Exercise::where('user_id', '=', Auth::id())
+       try
+       { 
+       		$exercise = Exercise::where('user_id', '=', Auth::id())
         	->findorfail($id);
-       }catch(Exception $e){
-       
+       }
+       catch(Exception $e)
+       {
 			$errorvalue = Helper::getErrorMessage($e);
 
 			return Redirect::to('/exercise/index')
 				->with('flash_message', 'Edit Failed...')
 				->withErrors($errorvalue);
        }
+        
         return View::make('exercise_edit')
         	->with('exercise', $exercise);
       
@@ -86,29 +85,25 @@ class ExerciseController extends BaseController
 
     public function postEdit()
     {
-    	try {
-    	# Step 1) Define the rules
-		$rules = array(
-			'desc' => 'required'
-		);
-
-		# Step 2)
-		$validator = Validator::make(Input::all(), $rules);
-
-		# Step 3
-		if($validator->fails()) {
-
-			return Redirect::to('/exercise/Edit')
-				->with('flash_message', 'Edit failed; Please try again.')
-				->withInput()
-				->withErrors($validator);
-		}
-
-    	
-
-	        $exercise = Exercise::findOrFail(Input::get('id'));
+    	try 
+    	{
+			$rules = array(
+				'desc' => 'required');
+	
+			$validator = Validator::make(Input::all(), $rules);
+	
+			if($validator->fails()) 
+			{
+				return Redirect::to('/exercise/Edit')
+					->with('flash_message', 'Edit failed; Please try again.')
+					->withInput()
+					->withErrors($validator);
+			}
+	
+		    $exercise = Exercise::findOrFail(Input::get('id'));
 	    }
-	    catch(exception $e) {
+	    catch(exception $e) 
+	    {
 	        $errorvalue = Helper::getErrorMessage($e);
 
 	        return Redirect::to('/exercise/index')
@@ -120,11 +115,12 @@ class ExerciseController extends BaseController
 	    
 		$exercise->desc = filter_var(Input::get('desc'), FILTER_SANITIZE_STRING);
 		
-		try {
+		try 
+		{
 			$exercise->save();
 		}
-		catch (Exception $e) {
-		
+		catch (Exception $e) 
+		{
 			$errorvalue = Helper::getErrorMessage($e);
 
 			return Redirect::to('/exercise/edit/'.Input::get('id'))
@@ -133,27 +129,42 @@ class ExerciseController extends BaseController
 
 		}
 
-		
 		return Redirect::to('/exercise/index');
 
     }
 
-    public function getDelete($id)
+    public function postDelete()
     {
-       try{ 
-       $exercise = Exercise::where('user_id', '=', Auth::id())
-        	->findorfail($id);
-        $exercise->delete();
-        	
-       }catch(Exception $e){
-       
-			$errorvalue = Helper::getErrorMessage($e);
+    	/* Ajax Delete */
+    	if(Request::ajax()) 
+    	{
+    		$eid  = Input::get('eid');
+    		try
+    		{
+	       		$exercise = Exercise::where('user_id', '=', Auth::id())
+	        	->findorfail($eid);
+	        	
+	        	$exercise->delete();
+	        	
+	        	return Response::json(array('success' => true), 200);
 
-			return Redirect::to('/exercise/index')
-				->with('flash_message', 'Delete Failed...')
-				->withErrors($errorvalue);
-       }
-       return Redirect::to('/exercise/index');
+			}
+			catch(Exception $e)
+			{
+				$errorvalue = Helper::getErrorMessage($e);
+				
+				return Response::json(array(
+		        'success' => false,
+		        'errors' => $errorvalue
+		    	), 400); 
+		    		
+	    	}
+	    	
+
+		}
+    		
+    		
+    		
     }
